commit b9

+++ module/restaurant/biz/delete_restaurant.go
@@ -3,7 +3,7 @@ package restaurantbiz
 import (
 	"context"
 	"errors"
-	"g05-fooddelivery/module/common"
+	"g05-fooddelivery/common"
 	restaurantmodel "g05-fooddelivery/module/restaurant/model"
 )
 
@@ -16,11 +16,12 @@ type DeleteRestaurantStore interface {
 	Delete(context context.Context, id int) error
 }
 type deleteRestaurantBiz struct {
-	store DeleteRestaurantStore
+	store     DeleteRestaurantStore
+	requester common.Requester
 }
 
-func NewDeleteRestaurantBiz(store DeleteRestaurantStore) *deleteRestaurantBiz {
-	return &deleteRestaurantBiz{store: store}
+func NewDeleteRestaurantBiz(store DeleteRestaurantStore, requester common.Requester) *deleteRestaurantBiz {
+	return &deleteRestaurantBiz{store: store, requester: requester}
 }
 func (biz *deleteRestaurantBiz) DeleteRestaurant(context context.Context, id int) error {
 	oldData, err := biz.store.FindDataWithCondition(context, map[string]interface{}{"id": id})
@@ -32,6 +33,9 @@ func (biz *deleteRestaurantBiz) DeleteRestaurant(context context.Context, id int
 	if oldData.Status == 0 {
 		//common.ErrEntityDeleted(restaurantmodel.EntityName,nil)
 		return errors.New("id empty")
+	}
+	if oldData.UserId != biz.requester.GetUserId() {
+		return common.ErrNoPermission(nil)
 	}
 	if err := biz.store.Delete(context, id); err != nil {
 		return common.ErrCannotDeleteEntity(restaurantmodel.EnityName, nil)
